Revert "remove cpp and merge in embree_params"
This reverts commit e040c93ceaf00f278acd035598d27b370304010f.
