+++ plugin/accel_embree/accel_embree.cpp
@@ -7,7 +7,7 @@
 #include <lm/scene.h>
 #include <lm/mesh.h>
 #include <lm/exception.h>
-#include "embree_params.h"
+#include "embree.h"
 
 LM_NAMESPACE_BEGIN(LM_NAMESPACE)
 
@@ -40,16 +40,19 @@ private:
 public:
 
      virtual void construct(const Json& prop) override {
-        settings_ = prop;
-        sf_ = prop;
+        settings_ = rtcDefaultBuildArguments();
+        from_json(prop,settings_);
 
-        //check actual values used for building
+        from_json(prop,sf_);
+
         Json j;
-        j = settings_;
+        to_json(j,settings_);
         LM_INFO(j.dump());
         j.clear();
-        j = sf_;
+        
+        to_json(j,sf_);
         LM_INFO( j.dump() );
+        
         }
 
     Accel_Embree() {
